v3.10.0 - 릴리즈 참고

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -877,16 +877,16 @@ public abstract class Mob extends Char {
 			}
 		}
 
+		if (this.alignment == Alignment.ENEMY && hero.hasTalent(Talent.BULLET_COLLECT)) {
+			BulletItem bullet = new BulletItem();
+			bullet.quantity(hero.pointsInTalent(Talent.BULLET_COLLECT));
+			Dungeon.level.drop(bullet, pos).sprite.drop();
+		}
+
 		if (cause == hero) {
 			if (Dungeon.hero.hasTalent(Talent.LETHAL_RAGE)){
 				Berserk berserk = Buff.affect(hero, Berserk.class);
 				berserk.add(0.067f*Dungeon.hero.pointsInTalent(Talent.LETHAL_RAGE));
-			}
-
-			if (hero.hasTalent(Talent.BULLET_COLLECT)) {
-				BulletItem bullet = new BulletItem();
-				bullet.quantity(hero.pointsInTalent(Talent.BULLET_COLLECT));
-				Dungeon.level.drop(bullet, pos).sprite.drop();
 			}
 
 			if (hero.hasTalent(Talent.SOUL_BULLET)) {
